requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
bcrypt==5.0.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
Django==5.2.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.4.0
passlib==1.7.4
pillow==12.1.0
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyJWT==2.10.1
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
requests==2.31.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
starlette==0.50.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.40.0
